+++ packages/math-input/src/data/key-configs.ts
@@ -193,6 +193,12 @@ const KeyConfigs = (
             ariaLabel: strings.cube,
         }),
     },
+    SUB: {
+        ...getDefaultOperatorFields({
+            key: "SUB",
+            ariaLabel: strings.subscript,
+        }),
+    },
     SQRT: {
         ...getDefaultOperatorFields({
             key: "SQRT",